+++ server/src/main/java/eu/openanalytics/phaedra/plateservice/service/PlateService.java
@@ -1,8 +1,9 @@
 package eu.openanalytics.phaedra.plateservice.service;
 
-import eu.openanalytics.phaedra.plateservice.model.Plate;
-import eu.openanalytics.phaedra.plateservice.repository.PlateRepository;
-import eu.openanalytics.phaedra.platservice.dto.PlateDTO;
+import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
+
 import org.modelmapper.Conditions;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.config.Configuration;
@@ -10,9 +11,11 @@ import org.modelmapper.convention.NameTransformers;
 import org.modelmapper.convention.NamingConventions;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
+import eu.openanalytics.phaedra.plateservice.model.Plate;
+import eu.openanalytics.phaedra.plateservice.repository.PlateRepository;
+import eu.openanalytics.phaedra.platservice.dto.ExperimentDTO;
+import eu.openanalytics.phaedra.platservice.dto.PlateDTO;
+import eu.openanalytics.phaedra.platservice.enumartion.ProjectAccessLevel;
 
 @Service
 public class PlateService {
@@ -20,10 +23,14 @@ public class PlateService {
 
 	private final PlateRepository plateRepository;
 	private final WellService wellService;
-
-	public PlateService(PlateRepository plateRepository, WellService wellService) {
+	private final ExperimentService experimentService;
+	private final ProjectAccessService projectAccessService;
+	
+	public PlateService(PlateRepository plateRepository, WellService wellService, ExperimentService experimentService, ProjectAccessService projectAccessService) {
 		this.plateRepository = plateRepository;
 		this.wellService = wellService;
+		this.experimentService = experimentService;
+		this.projectAccessService = projectAccessService;
 
 		// TODO move to dedicated ModelMapper service
 		Configuration builderConfiguration = modelMapper.getConfiguration().copy()
@@ -34,6 +41,13 @@ public class PlateService {
 	}
 
 	public PlateDTO createPlate(PlateDTO plateDTO) {
+		long projectId = Optional.ofNullable(experimentService.getExperimentById(plateDTO.getExperimentId()))
+				.map(ExperimentDTO::getProjectId)
+				.orElse(0L);
+		if (projectId == 0L) return null; // Experiment not found or not accessible.
+		
+		projectAccessService.checkAccessLevel(projectId, ProjectAccessLevel.Write);
+		
 		Plate plate = new Plate();
 		modelMapper.typeMap(PlateDTO.class, Plate.class)
 				.map(plateDTO, plate);
@@ -48,6 +62,7 @@ public class PlateService {
 	public PlateDTO updatePlate(PlateDTO plateDTO) {
 		Optional<Plate> plate = plateRepository.findById(plateDTO.getId());
 		plate.ifPresent(p -> {
+			projectAccessService.checkAccessLevel(getProjectIdByPlateId(p.getId()), ProjectAccessLevel.Write);
 			modelMapper.typeMap(PlateDTO.class, Plate.class)
 					.setPropertyCondition(Conditions.isNotNull())
 					.map(plateDTO, p);
@@ -57,27 +72,44 @@ public class PlateService {
 	}
 
 	public void deletePlate(long plateId) {
+		projectAccessService.checkAccessLevel(getProjectIdByPlateId(plateId), ProjectAccessLevel.Write);
 		plateRepository.deleteById(plateId);
 	}
 
 	public List<PlateDTO> getAllPlates() {
 		List<Plate> result = (List<Plate>) plateRepository.findAll();
-		return result.stream().map(this::mapToPlateDTO).collect(Collectors.toList());
+		return result.stream()
+				.filter(p -> projectAccessService.hasAccessLevel(getProjectIdByPlateId(p.getId()), ProjectAccessLevel.Read))
+				.map(this::mapToPlateDTO)
+				.toList();
 	}
 
 	public List<PlateDTO> getPlatesByExperimentId(long experimentId) {
 		List<Plate> result = plateRepository.findByExperimentId(experimentId);
-		return result.stream().map(this::mapToPlateDTO).collect(Collectors.toList());
+		return result.stream()
+				.filter(p -> projectAccessService.hasAccessLevel(getProjectIdByPlateId(p.getId()), ProjectAccessLevel.Read))
+				.map(this::mapToPlateDTO)
+				.toList();
 	}
 
 	public List<PlateDTO> getPlatesByBarcode(String barcode) {
 		List<Plate> result = plateRepository.findByBarcode(barcode);
-		return result.stream().map(this::mapToPlateDTO).collect(Collectors.toList());
+		return result.stream()
+				.filter(p -> projectAccessService.hasAccessLevel(getProjectIdByPlateId(p.getId()), ProjectAccessLevel.Read))
+				.map(this::mapToPlateDTO)
+				.toList();
 	}
 
 	public PlateDTO getPlateById(long plateId) {
 		Optional<Plate> result = plateRepository.findById(plateId);
-		return result.map(this::mapToPlateDTO).orElse(null);
+		return result
+				.filter(p -> projectAccessService.hasAccessLevel(getProjectIdByPlateId(p.getId()), ProjectAccessLevel.Read))
+				.map(this::mapToPlateDTO)
+				.orElse(null);
+	}
+	
+	private Long getProjectIdByPlateId(long plateId) {
+		return plateRepository.findProjectIdByPlateId(plateId);
 	}
 
 	private PlateDTO mapToPlateDTO(Plate plate) {
